requirements.txt
Django==4.2.4
djangorestframework==3.14.0
Jinja2==3.1.3
Terraform==0.10.1
